refactor: 분석 데이터 가공 로직을 공통 유틸로 분리 #99
