+++ fe/src/hooks/useAnalysisData.ts
@@ -1,8 +1,4 @@
-import {
-  CategoryAnalysis,
-  TransactionAnalysis,
-  TransactionType,
-} from '@/types/analysis';
+import { TransactionAnalysis, TransactionType } from '@/types/analysis';
 import { formatLocalDate, getMonthRange, minDate } from '@/utils/date';
 import { useEffect, useMemo, useState } from 'react';
 
@@ -10,15 +6,11 @@ import { fetchTransactionByRange } from '@/services/analysis/analysisService';
 import { getCurrentUser } from '@/services/analysis/budgetService';
 import { syncByMonthClient } from '@/services/fixed-costs/syncFixedTransactions.client';
 import { toast } from 'sonner';
+import { transformAnalysisData } from '@/utils/analysis-transform';
 
 type AnalysisState = {
   current: TransactionAnalysis[];
   prev: TransactionAnalysis[];
-};
-
-// 카테고리 명: 값(합계 금액)
-type CategoryGroup = {
-  [key: string]: number;
 };
 
 const safeSyncTransactions = async (date: Date, through: string) => {
@@ -90,62 +82,16 @@ export const useAnalysisData = (selectedDate: Date, type: TransactionType) => {
     fetchData();
   }, [selectedDate, type]);
 
-  const { totalAmount, prevAmount, categoryData, categoryTotalsByKey } =
-    useMemo(() => {
-      // 현재/이전 달 총액 계산
-      const currentTotal = current.reduce(
-        (sum, item) => sum + (item.amount || 0),
-        0
-      );
-      const lastTotal = prev.reduce((sum, item) => sum + (item.amount || 0), 0);
-
-      // 카테고리 ID별 합계를 담을 객체
-      const totalsByKey: Record<string, number> = {};
-
-      // 카테고리별 그룹화 및 합계 계산
-      const grouped = current.reduce<CategoryGroup>((acc, item) => {
-        const categoryName = item.category?.name_ko || '기타';
-        const categoryKey = item.category?.category_key;
-
-        // 카테고리가 첫 등장이면 0으로 초기화
-        if (!acc[categoryName]) acc[categoryName] = 0;
-        // 거래 금액 합산 (누적)
-        acc[categoryName] += item.amount || 0;
-
-        if (categoryKey) {
-          totalsByKey[categoryKey] =
-            (totalsByKey[categoryKey] || 0) + (item.amount || 0);
-        }
-
-        return acc;
-      }, {});
-
-      // 배열 변환 및 정렬, 비율 계산
-      const sortedCategoryData: CategoryAnalysis[] = Object.entries(grouped)
-        .map(([name, amount]) => ({
-          name,
-          amount,
-          percentage: currentTotal > 0 ? (amount / currentTotal) * 100 : 0,
-        }))
-        .sort((a, b) => b.amount - a.amount);
-
-      return {
-        totalAmount: currentTotal,
-        prevAmount: lastTotal,
-        categoryData: sortedCategoryData,
-        categoryTotalsByKey: totalsByKey,
-      };
-    }, [current, prev]);
-
-  const diff = totalAmount - prevAmount;
+  // 공통 가공 함수 호출
+  const analysisResult = useMemo(
+    () => transformAnalysisData(data.current, data.prev),
+    [data.current, data.prev]
+  );
 
   return {
     current: current,
     prev: prev,
-    totalAmount,
-    diff,
     isLoading,
-    categoryData,
-    categoryTotalsByKey,
+    ...analysisResult,
   };
 };
